[DM-23286] Update nginx TLS configuration settings
The previous settings received only a B grade from SSL Labs
because they allowed TLS 1.0 and TLS 1.1 and allowed some older
ciphers that are no longer considered secure.

The new settings are based on the Mozilla SSL Configuration
Generator, which is now the recommended resource for strong SSL
configuration, using the intermediate settings option.

Update README.md accordingly with a pointer to the configuration
generator.